Allow plain http connections when skipTLS is configured
When communicating with a registry that uses plain http, but is not detectable as being on a loopback address, such as localhost or 127.x.x.x, the calls to the registry fail because we try to use https the registry.

I have added a workaround so that when the caller provides a custom http transport, and it has TLS verification disabled, we should also consider http as an allowed protocol for communicating with the registry. This fit the most closely to the original call patterns in the image-relocation library, and is a backwards-compatible change (in that the code will compile against this change, obviously the behavior is different).

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>